jwt-security-tools/requirements.txt
PyJWT
requests
flask
pytest